+++ server/controllers/ExampleController.js
@@ -1,4 +1,4 @@
-const exampleModel = require('../models/example');
+const ExampleModel = require('../models/ExampleModel');
 const debug = require('../utils/debug');
 
 module.exports = {
@@ -8,8 +8,7 @@ module.exports = {
 
   postHandler(req, res) {
     debug.logRequest(req);
-    exampleModel
-      .findOneAndUpdate(req.query, req.body, { upsert: true })
+    ExampleModel.findOneAndUpdate(req.query, req.body, { upsert: true })
       .exec()
       .then((result) => {
         // console.log('result: ', result);
